Properly add doi_evictions to dagster dag
